# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Recruitment ATS + CRM API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /positions:
    get:
      operationId: listPositions
      tags: [positions]
      summary: List positions
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: clientId
          in: query
          schema:
            type: integer
        - name: location
          in: query
          schema:
            type: string
        - name: priority
          in: query
          schema:
            type: string
        - name: employmentType
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Positions list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Position"
    post:
      operationId: createPosition
      tags: [positions]
      summary: Create position
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PositionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Position"
  /positions/{id}:
    get:
      operationId: getPosition
      tags: [positions]
      summary: Get position
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Position with candidates
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PositionDetail"
    patch:
      operationId: updatePosition
      tags: [positions]
      summary: Update position
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PositionUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Position"
    delete:
      operationId: deletePosition
      tags: [positions]
      summary: Delete position
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── CANDIDATES ───────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Position:
      type: object
      required: [id, clientId, clientName, positionName, location, employmentType, priority, openings, status, createdAt]
      properties:
        id:
          type: integer
        clientId:
          type: integer
        clientName:
          type: string
        positionName:
          type: string
        jobDescription:
          type: ["string", "null"]
        jobDescriptionLink:
          type: ["string", "null"]
        location:
          type: string
        employmentType:
          type: string
        priority:
          type: string
        openings:
          type: integer
        hiringManager:
          type: ["string", "null"]
        status:
          type: string
        createdAt:
          type: string
        candidateCount:
          type: ["integer", "null"]
    PositionDetail:
      type: object
      required: [id, clientId, clientName, positionName, location, employmentType, priority, openings, status, createdAt, candidates]
      properties:
        id:
          type: integer
        clientId:
          type: integer
        clientName:
          type: string
        positionName:
          type: string
        jobDescription:
          type: ["string", "null"]
        jobDescriptionLink:
          type: ["string", "null"]
        location:
          type: string
        employmentType:
          type: string
        priority:
          type: string
        openings:
          type: integer
        hiringManager:
          type: ["string", "null"]
        status:
          type: string
        createdAt:
          type: string
        candidates:
          type: array
          items:
            $ref: "#/components/schemas/Candidate"
    PositionInput:
      type: object
      required: [clientId, positionName, location, employmentType, priority, openings]
      properties:
        clientId:
          type: integer
        positionName:
          type: string
        jobDescription:
          type: string
        jobDescriptionLink:
          type: string
        location:
          type: string
        employmentType:
          type: string
        priority:
          type: string
        openings:
          type: integer
        hiringManager:
          type: string
    PositionUpdate:
      type: object
      properties:
        positionName:
          type: string
        jobDescription:
          type: string
        jobDescriptionLink:
          type: string
        location:
          type: string
        employmentType:
          type: string
        priority:
          type: string
        openings:
          type: integer
        hiringManager:
          type: string
        status:
          type: string

    # ─── CANDIDATE ─────────────────────────────────────────────────────────────
    Candidate:
      type: object
      required: [id, positionId, positionName, clientId, clientName, candidateName, email, phone, status, source, submissionDate, createdAt]
      properties:
        id:
          type: integer
        positionId:
          type: integer
        positionName:
          type: string
        clientId:
          type: integer
        clientName:
          type: string
        candidateName:
          type: string
        email:
          type: string
        phone:
          type: string
        currentCompany:
          type: ["string", "null"]
        experience:
          type: ["string", "null"]
        currentCtc:
          type: ["string", "null"]
        expectedCtc:
          type: ["string", "null"]
        noticePeriod:
          type: ["string", "null"]
        source:
          type: string
        status:
          type: string
        submissionDate:
          type: string
        createdAt:
          type: string
        recruiterId:
          type: ["integer", "null"]
        recruiterName:
          type: ["string", "null"]
        resumeFileName:
          type: ["string", "null"]
...