# lib/api-zod/src/generated/types/positionInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Recruitment ATS + CRM API
 * OpenAPI spec version: 0.1.0
 */

export interface PositionInput {
  clientId: number;
  positionName: string;
  jobDescription?: string;
  jobDescriptionLink?: string;
  location: string;
  employmentType: string;
  priority: string;
  openings: number;
  hiringManager?: string;
}